refactor: store labels as u16 instead of u32

+++ src/inertia.rs
@@ -1,7 +1,7 @@
 use crate::distance::{earth_movers_distance, euclidian_distance};
 
 
-pub fn calculate_inertia_emd(data: &[u8], histogram_size: usize, centroids: &[Vec<f64>], labels: &[u32]) -> f64 {
+pub fn calculate_inertia_emd(data: &[u8], histogram_size: usize, centroids: &[Vec<f64>], labels: &[u16]) -> f64 {
     let result = labels.iter().enumerate().map(|(i, &label)| {
         let point_start = i * histogram_size;
         let point = &data[point_start..point_start + histogram_size];
@@ -16,7 +16,7 @@ pub fn calculate_inertia_emd(data: &[u8], histogram_size: usize, centroids: &[Ve
     result
 }
 
-pub fn calculate_inertia_emd_f64(data: &[f64], histogram_size: usize, centroids: &[Vec<f64>], labels: &[u32]) -> f64 {
+pub fn calculate_inertia_emd_f64(data: &[f64], histogram_size: usize, centroids: &[Vec<f64>], labels: &[u16]) -> f64 {
     let result = labels.iter().enumerate().map(|(i, &label)| {
         let point_start = i * histogram_size;
         let point = &data[point_start..point_start + histogram_size].to_vec();
@@ -31,7 +31,7 @@ pub fn calculate_inertia_emd_f64(data: &[f64], histogram_size: usize, centroids:
     result
 }
 
-pub fn calculate_inertia_euclidian(data: &[u8], histogram_size: usize, centroids: &[Vec<f64>], labels: &[u32]) -> f64 {
+pub fn calculate_inertia_euclidian(data: &[u8], histogram_size: usize, centroids: &[Vec<f64>], labels: &[u16]) -> f64 {
     let result = labels.iter().enumerate().map(|(i, &label)| {
         let point_start = i * histogram_size;
         let point = &data[point_start..point_start + histogram_size];
